fix: update file structure
