+++ pkg/coingecko/svc/get_coin.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 	db "github.com/markpassawat/go-grpc-coinlist/pkg/coingecko/db"
-	pb "github.com/markpassawat/go-grpc-coinlist/pkg/coingecko/route"
+	pb "github.com/markpassawat/go-grpc-coinlist/proto/coinlist"
 )
 
 func (s *CoinServer) GetCoin(ctx context.Context,
